pyproject.toml
[project]
name = "finbuddy-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "groq>=0.29.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uagents>=0.22.5",
    "uvicorn[standard]>=0.35.0",
]
